double checked weight works.Need to do the heap

+++ kernel/kernel.h
@@ -45,7 +45,7 @@ typedef struct {
 typedef struct {
   uint32_t pc;
   uint8_t active;
-  uint16_t wt;
+  uint32_t wt;
 } entry_t;
 
 extern uint32_t boh;
